requirements.txt
beautifulsoup4==4.13.3
certifi==2025.1.31
cryptography==42.0.5
easy_vault==0.7.0
humanize==4.12.2
lxml==5.2.2
mss==9.0.1
openpyxl==3.1.5
pandas==2.2.3
paramiko==3.4.0
playwright==1.42.0
psutil==5.9.8
python_dateutil==2.8.2
pywin32==306
Requests==2.32.3
retrying==1.3.4
tinydb==4.8.2
urllib3==1.26.14
wrapt_timeout_decorator==1.5.1
spacy==3.8.7
lmdb==1.7.3

